feat(ai): implement board evaluation, lookahead targeting, piece generation, and scoring strategies for AI player
